requirements.txt
Django==2.2.7
django-cors-headers==3.0.2
django-environ==0.4.5
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
geopy==1.20.0
psycopg2-binary==2.8.3